Did a whole shitload of piss
Well I didn't technically do that much but it definitely took like 8 hours. I fixed the god damn 3 problem, which was caused by mutation errors. For some reason using Arrays.copyOf() wouldn't fix the problem. No idea why. But anyway, writing my own .copyOf() fixed it, somehow. The MCTS actually runs and makes decisions, but the decisions are still random due to the simulation method always returning the same result. It does this whether it's my simulation method or the one Nick wrote, for some reason, except Nick's always returns 0 and mine always returns either 1 or -1 depending on which team the AI is on. So, we'll have to fix that and then we might have a functioning AI.