ITS WORKING NOW
Yep, a lot of things were changed, but they were mostly small details. It's possible that what has it working properly now, intelligently, is that the UCB calculation was broken but has been fixed. The AI manages to survive as long as physically possible in some games (until the entire board is filled with arrows) when playing against itself.